requirements.txt
flask>=2.0.0
face-recognition>=1.3.0
numpy>=1.21.0
pillow>=8.0.0
opencv-python>=4.5.0
scipy>=1.7.0
python-dotenv>=0.19.0
boto3>=1.26.0  # Only if you're using AWS features
